refactor: resetElements와 constructor 역할 구분
